Fix ForwardingInformation unmarshaling (TAPI <3.1)
In older versions of TAPI dwCallerAddressType and dwDestAddressType were not present in the LINEFORWARD struct. Therefore only 24 instead of 32 bytes are transmitted for every entry of the array.

To fix this issue I only copy as many bytes as the defined in the LINEADDRESSSTATUS dwForwardSize field. This should work for all versions of TAPI.